% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppADutils-tests}
\alias{CppADutils-tests}
\alias{CppADutils_tests}
\alias{cppad_lbeta1}
\alias{cppad_dnorm_log}
\alias{cppad_dbeta_log}
\alias{cppad_dlogitbeta_log}
\alias{cppad_lgamma1}
\alias{cppad_lgamma1p}
\alias{cppad_lgammaexp}
\alias{cppad_log1p}
\alias{cppad_expm1}
\alias{cppad_log1pexp}
\alias{cppad_log1pmx}
\alias{cppad_loginvlogit}
\alias{cppad_invlogit}
\alias{cppad_incgamma}
\alias{cppad_incbeta}
\alias{cppad_incbeta2}
\alias{cppad_lgammaLogExp}
\alias{cppad_dt_log}
\alias{cppad_dhalft_log}
\alias{cppad_dhalft_log2}
\alias{cppad_dnormTrunc0_log}
\alias{cppad_atan2a}
\alias{cppad_atan2b}
\alias{cppad_pnorm_log}
\alias{cppad_pnorm}
\alias{cppad_dgamma_log}
\title{tests}
\usage{
CppADutils_tests(X)

cppad_lbeta1(X)

cppad_dnorm_log(X)

cppad_dbeta_log(X)

cppad_dlogitbeta_log(X)

cppad_lgamma1(X)

cppad_lgamma1p(X)

cppad_lgammaexp(X)

cppad_log1p(X)

cppad_expm1(X)

cppad_log1pexp(X)

cppad_log1pmx(X)

cppad_loginvlogit(X)

cppad_invlogit(X)

cppad_incgamma(X)

cppad_incbeta(X)

cppad_incbeta2(X)

cppad_lgammaLogExp(X)

cppad_dt_log(X)

cppad_dhalft_log(X)

cppad_dhalft_log2(X)

cppad_dnormTrunc0_log(X)

cppad_atan2a(X)

cppad_atan2b(X)

cppad_pnorm_log(X)

cppad_pnorm(X)

cppad_dgamma_log(X)
}
\arguments{
\item{X}{numeric vector (not matrix at this time)}
}
\description{
tests
}
