% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dgamma_test}
\alias{dgamma_test}
\title{gamma_logpdf test}
\usage{
dgamma_test(X_, R_, A_)
}
\arguments{
\item{X_}{vector}

\item{R_}{vector}

\item{A_}{vector}
}
\value{
Numeric vector
}
\description{
gamma_logpdf test
}
