% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgamma1p}
\alias{lgamma1p}
\title{log(gamma(1+x))}
\usage{
lgamma1p(X)
}
\arguments{
\item{X}{input numeric vector}
}
\value{
log(gamma(1+x))
}
\description{
log(gamma(1+x)), accurate even for very small |x|
}
