% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgammaexp}
\alias{lgammaexp}
\title{log(gamma(exp(x)))}
\usage{
lgammaexp(X)
}
\arguments{
  \item{X}{input numeric vector}
}
\value{
log(gamma(exp(x)))
}
\description{
log(gamma(exp(x))), accurate even for very small |x|
}

