% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LKJ}
\alias{LKJ}
\title{LKJ pdf}
\usage{
LKJ(Y_, eta_, K)
}
\arguments{
  \item{Y_}{unconstrained vector} \item{eta_}{parameter >0}
  \item{K}{dimension integer}
}
\value{
log pdf
}
\description{
LKJ pdf
}

