% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LKJ_const}
\alias{LKJ_const}
\title{LKJ constant term in pdf}
\usage{
LKJ_const(eta, K)
}
\arguments{
  \item{K}{dimension integer} \item{eta_}{parameter >0}
}
\value{
log pdf
}
\description{
LKJ constant term in pdf
}

