% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LKJ_test}
\alias{LKJ_test}
\title{LKJ test}
\usage{
LKJ_test(Y_, eta_, K)
}
\arguments{
\item{Y_}{unconstrained vector}

\item{eta_}{parameter >0}

\item{K}{dimension integer}
}
\value{
List
}
\description{
LKJ test
}
