% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LKJ_unwrap}
\alias{LKJ_unwrap}
\title{Unwrap LKJ}
\usage{
LKJ_unwrap(Y_, K)
}
\arguments{
  \item{K}{integer dimension} \item{Y}{input numeric vector}
}
\value{
matrix
}
\description{
Unwrap LKJ
}

