% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log1pexp}
\alias{log1pexp}
\title{log(1+exp(x))}
\usage{
log1pexp(X)
}
\arguments{
  \item{X}{input numeric vector}
}
\value{
log(1+exp(x))
}
\description{
log(1+exp(x)), accurate even for very large X.
}

