% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVN_test}
\alias{MVN_test}
\title{MVN test}
\usage{
MVN_test(X_, mu_, G_, isPrec)
}
\arguments{
  \item{X_}{matrix} \item{mu_}{matrix} \item{G_}{full
  covariance or precision matrix} \item{isPrec}{covariance or
  precision matrix?}
}
\value{
Numeric vector
}

