% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppADutils-class.R
\docType{class}
\name{adtest-class}
\alias{adtest-class}
\alias{ads}
\alias{Rcpp_adtest-class}
\title{AD test for CppADutils}
\format{An object of class \code{character} of length 1.}
\usage{
ads
}
\description{
class to run tests
}
\keyword{datasets}
