% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CppADutils-class.R
\docType{class}
\name{adtest-class}
\alias{Rcpp_adtest-class}
\alias{ads}
\alias{adtest-class}
\title{AD test for CppADutils}
\format{\preformatted{ chr ".1"
}}
\usage{
ads
}
\description{
class to run tests
}
\keyword{datasets}

