% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loginvlogit}
\alias{loginvlogit}
\title{Log inverse logit function}
\usage{
loginvlogit(X)
}
\arguments{
  \item{X}{input numeric vector}
}
\value{
p = log(exp(x)/(1+exp(x)) = x-log(1+exp(x))
}
\description{
Log inverse logit, accurate even for very negative x
}

